pyproject.toml
[project]
name = "discord-messages-relay"
version = "0.1.0"
description = "Discord message relay server"
requires-python = ">=3.12"
authors = [{name = "Relay"}]
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.24",
  "sqlalchemy>=2.0",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "pyyaml>=6.0",
  "discord.py>=2.4",
  "httpx>=0.26",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "anyio>=4.0"
]

[project.scripts]
relayctl = "relay_client.cli:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
